coverity: minor fixes
Fix a few minor bugs.

+++ pgppacket.c
@@ -138,7 +138,7 @@ unsigned char *pgp_read_packet (FILE * fp, size_t * len)
 	  perror ("fread");
 	  goto bail;
 	}
-	material = buf[0] << 24;
+	material = (size_t) buf[0] << 24;
 	material |= buf[1] << 16;
 	material |= buf[2] << 8;
 	material |= buf[3];
